Reparat locatia fisierelor
